+++ components/Question.js
@@ -20,6 +20,7 @@ class Question extends React.Component {
       askedBy,
       isUserCourseStaff,
       isUserAnsweringQuestion,
+      isUserAnsweringOtherQuestion,
       didUserAskQuestion,
     } = this.props
 
@@ -58,16 +59,17 @@ class Question extends React.Component {
     } else {
       buttonCluster = (
         <Fragment>
-          {isUserCourseStaff && (
-            <Button
-              color="primary"
-              outline
-              className="mr-2"
-              onClick={() => this.props.startQuestion()}
-            >
-              Start Answering!
-            </Button>
-          )}
+          {isUserCourseStaff &&
+            !isUserAnsweringOtherQuestion && (
+              <Button
+                color="primary"
+                outline
+                className="mr-2"
+                onClick={() => this.props.startQuestion()}
+              >
+                Start Answering!
+              </Button>
+            )}
           {didUserAskQuestion && (
             <Button
               color="primary"
@@ -178,6 +180,7 @@ Question.propTypes = {
   didUserAskQuestion: PropTypes.bool.isRequired,
   isUserCourseStaff: PropTypes.bool.isRequired,
   isUserAnsweringQuestion: PropTypes.bool.isRequired,
+  isUserAnsweringOtherQuestion: PropTypes.bool.isRequired,
   cancelQuestion: PropTypes.func.isRequired,
   startQuestion: PropTypes.func.isRequired,
   editQuestion: PropTypes.func.isRequired,